requirements.txt
# Core requirements
streamlit
langchain
langchain-google-genai
google-generativeai
langchain-community
langchain-core
langchain-google-genai
# PDF processing
PyPDF2
# Vector database
faiss-cpu

# Database
psycopg2-binary

# Environment management
python-dotenv

# Additional utilities
numpy
pandas
typing-extensions